changing from getting all titles and assuming a row order to parsing a full spell block/all spell paragraphs just below the target title.
I didnot parse schools/descriptors/levelsPerClass for 3 spells correctly before,
so I have to adjust the expected results. The 3 spells were:
-http://paizo.com/pathfinderRPG/prd/coreRulebook/spells/summonMonster.html#summon-monster-i (parsed ii)
-http://paizo.com/pathfinderRPG/prd/coreRulebook/spells/summonNatureSAlly.html#summon-nature-s-ally-i (parsed ii)
-http://paizo.com/pathfinderRPG/prd/ultimateMagic/spells/undeadAnatomy.html#undead-anatomy-iv (parsed i)

However, I have taken away the possibility to fallback for other titles on the same page.
Therefore, the tests fail for
http://paizo.com/pathfinderRPG/prd/advancedPlayersGuide/spells/planarAdaptation.html#planar-adaptation,-mass
because it has no school. I would like to implement a recursive spell parser to parse the description for other spells.